Modeling of methods on boxed primitive classes
Bug: b/541236096

+++ src/main/java/com/android/tools/r8/ir/analysis/value/AbstractValueFactory.java
@@ -39,7 +39,7 @@ public class AbstractValueFactory {
     return SingleBoxedBooleanValue.getTrueInstance();
   }
 
-  public SingleBoxedByteValue createBoxedByte(int value) {
+  public SingleBoxedByteValue createBoxedByte(byte value) {
     return new SingleBoxedByteValue(value);
   }
